backend/requirements.txt
# Pinned versions for reproducible builds
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.17
slowapi==0.1.9
python-dotenv==1.0.1
google-genai>=1.0.0
# Auth
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
# HTTP client (OAuth + general)
httpx==0.27.0
# PDF parsing
PyMuPDF==1.25.3
# Migrations
alembic==1.14.1
# Testing
pytest==8.3.4
